+++ src/main/java/andromeda/hebat/finalisjtiadmin/controllers/admin/jurusan/KelolaDataMahasiswaController.java
@@ -1,6 +1,6 @@
-package andromeda.hebat.finalisjtiadmin.controllers.jurusan;
+package andromeda.hebat.finalisjtiadmin.controllers.admin.jurusan;
 
-import andromeda.hebat.finalisjtiadmin.controllers.jurusan.overlay.OverlayEditDataMahasiswa;
+import andromeda.hebat.finalisjtiadmin.controllers.admin.jurusan.overlay.OverlayEditDataMahasiswa;
 import andromeda.hebat.finalisjtiadmin.Main;
 import andromeda.hebat.finalisjtiadmin.core.Database;
 import andromeda.hebat.finalisjtiadmin.models.Mahasiswa;
