+++ src/main/java/andromeda/hebat/finalisjtiadmin/controllers/admin/prodi/DashboardController.java
@@ -1,0 +1,11 @@
+package andromeda.hebat.finalisjtiadmin.controllers.admin.prodi;
+
+import andromeda.hebat.finalisjtiadmin.models.BerkasPengajuan;
+import javafx.fxml.FXML;
+import javafx.scene.control.TableView;
+
+public class DashboardController {
+    @FXML private TableView<BerkasPengajuan> tableViewDashboard;
+
+
+}
