+++ src/main/java/andromeda/hebat/finalisjtiadmin/controllers/admin/prodi/PermintaanVerifAllController.java
@@ -1,0 +1,5 @@
+package andromeda.hebat.finalisjtiadmin.controllers.admin.prodi;
+
+public class PermintaanVerifAllController {
+    
+}
