+++ src/main/java/andromeda/hebat/finalisjtiadmin/controllers/admin/ta/DashboardController.java
@@ -1,0 +1,4 @@
+package andromeda.hebat.finalisjtiadmin.controllers.admin.ta;
+
+public class DashboardController {
+}
