--- src/main/java/andromeda/hebat/finalisjtiadmin/controllers/prodi/DashboardController.java
@@ -1,5 +1,0 @@
-package andromeda.hebat.finalisjtiadmin.controllers.prodi;
-
-public class DashboardController {
-
-}
