--- src/main/java/andromeda/hebat/finalisjtiadmin/controllers/prodi/PermintaanVerifAllController.java
@@ -1,5 +1,0 @@
-package andromeda.hebat.finalisjtiadmin.controllers.prodi;
-
-public class PermintaanVerifAllController {
-    
-}
