--- src/main/java/andromeda/hebat/finalisjtiadmin/controllers/ta/DashboardController.java
@@ -1,4 +1,0 @@
-package andromeda.hebat.finalisjtiadmin.controllers.ta;
-
-public class DashboardController {
-}
